carbon_platform/services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

SQLAlchemy==2.0.36
asyncpg==0.30.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.12

orjson==3.10.7

aiofiles==24.1.0

httpx==0.27.2

boto3==1.34.162
botocore==1.34.162
